backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
python-dotenv==1.0.1
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai==0.3.2 
bcrypt 
cryptography


